refactor: extract project-image.js from project.js (PR-07)
Move hydrateImageRefs and saveImageFile into a dedicated module along
with the image directory setup. project.js reduced from 3,497 to
3,432 lines.
